+++ bricks/forms/src/user-or-user-group-select/index.tsx
@@ -11,7 +11,7 @@ import {
   UserOrUserGroupSelectValue,
 } from "./UserOrUserGroupSelect";
 import { CmdbModels } from "@next-sdk/cmdb-sdk";
-import { keyBy, groupBy, startsWith } from "lodash";
+import { groupBy, startsWith } from "lodash";
 import { FormItemElement } from "@next-libs/forms";
 
 /**
@@ -32,157 +32,131 @@ import { FormItemElement } from "@next-libs/forms";
  */
 export class UserOrUserGroupSelectElement extends FormItemElement {
   /**
-   * @kind string
    * @required true
    * @default -
    * @description 下拉框字段名
+   * @group basicFormItem
    */
   @property({ attribute: false }) declare name: string;
 
   /**
-   * @kind string
    * @required false
-   * @default -
    * @description 下拉框字段说明
+   * @group basicFormItem
    */
   @property({ attribute: false }) declare label: string;
 
   /**
-   * @kind {selectedUser: string[],selectedUserGroup: string[]}|string[]
-   * @required false
-   * @default -
    * @description 用户（组）选择构件中下拉框的初始值，按照我们平台的用户（组）数据，selectedUser 为"USER"模型中的 name，selectedUserGroup 为"USER_GROUP"模型中的":"+instanceId。当`mergeUseAndUserGroup`为 true 时，类型为`string[]`。
+   * @group basicFormItem
    */
   @property({
     attribute: false,
   })
-  value: string[] | UserOrUserGroupSelectValue;
+  value?: string[] | UserOrUserGroupSelectValue;
 
   /**
-   * @kind string
    * @required false
-   * @default -
    * @description 下拉框占位说明
+   * @group basicFormItem
    */
   @property({ attribute: false }) declare placeholder: string;
 
   /**
-   * @kind boolean
    * @required false
-   * @default -
    * @description 是否必填项
+   * @group basicFormItem
    */
   @property({ type: Boolean }) declare required: boolean;
 
   /**
-   * @kind boolean
-   * @required false
    * @default false
    * @description 是否隐藏“快速选择我”按钮
+   * @group ui
    */
   @property({
     type: Boolean,
   })
-  hideAddMeQuickly: boolean;
+  hideAddMeQuickly?: boolean;
 
   /**
-   * @kind string[]
-   * @required false
-   * @default -
    * @description 固定白名单列表，该列表中的值用户不能取消。
+   * @group basic
    */
   @property({
     attribute: false,
   })
-  staticList: string[];
+  staticList?: string[];
 
   /**
-   * @kind boolean
-   * @required false
    * @default false
    * @description 是否合并用户和用户组数据，当设置为 true 时，输入的`value`和`user.group.change`事件输出的 detail 都为`string[]`格式。
+   * @group basic
    */
   @property({
     type: Boolean,
   })
-  mergeUseAndUserGroup: boolean;
+  mergeUseAndUserGroup?: boolean;
 
   /**
-   * @kind "all"|"group"|"user"
-   * @required false
    * @default "all"
    * @description 支持选择用户、用户组或者两者
+   * @group ui
    */
   @property({ attribute: false })
-  optionsMode: "user" | "group" | "all" = "all";
+  optionsMode?: "user" | "group" | "all" = "all";
 
   /**
-   * @kind boolean
-   * @required false
    * @default false
    * @description 隐藏无效用户
-   * @group advanced
+   * @group ui
    */
   @property({ type: Boolean })
-  hideInvalidUser: boolean;
+  hideInvalidUser?: boolean;
 
   /**
-   * @kind boolean
-   * @required false
    * @default false
    * @description 是否隐藏搜索 icon，即不支持通过 cmdb 的 modal 选择器选择
-   * @group advanced
+   * @group ui
    */
   @property({
     type: Boolean,
   })
-  hideSelectByCMDB: boolean;
+  hideSelectByCMDB?: boolean;
 
   /**
-   * @kind `Record<string, any>`
-   * @required false
-   * @default -
    * @description  用户和用户组`search`接口的`query`，此参数比较适用于，两者接口需要参数相同的情况下使用
-   * @group advanced
+   * @group other
    */
   @property({
     attribute: false,
   })
-  query: Record<string, any>;
+  query?: Record<string, any>;
 
   /**
-   * @kind CmdbModels.ModelCmdbObject
-   * @required true
-   * @default -
-   * @description 模型列表，直接来自"providers-of-cmdb.cmdb-object-api-get-object-all"
-   * @group advanced
+   * @description 模型列表，不传该属性构件内部会发请求获取该列表，如果需要传该属性则优先使用外部传进来的数据，该数据来自"providers-of-cmdb.cmdb-object-api-get-object-ref" 如 demo 所示
+   * @group other
    */
   @property({ attribute: false })
-  objectList: Partial<CmdbModels.ModelCmdbObject>[];
+  objectList?: Partial<CmdbModels.ModelCmdbObject>[];
 
   /**
-   * @kind `Record<string, any>`
-   * @required false
-   * @default -
    * @description 针对`USER/instance/_search`接口的`query`，此参数比较适用于，可能只需要针对用户做筛选的情况
-   * @group advanced
+   * @group other
    */
   @property({
     attribute: false,
   })
-  userQuery: Record<string, any>;
+  userQuery?: Record<string, any>;
 
   /**
-   * @kind `Record<string, any>`
-   * @required false
-   * @default -
    * @description 针对`USER_GROUP/instance/_search`接口的`query`，此参数比较适用于，可能只需要针对用户组做筛选的情况
-   * @group advanced
+   * @group other
    */
   @property({
     attribute: false,
   })
-  userGroupQuery: Record<string, any>;
+  userGroupQuery?: Record<string, any>;
 
   /**
    * @detail `string[]|{selectedUser: string[],selectedUserGroup: string[]}`
@@ -216,8 +190,7 @@ export class UserOrUserGroupSelectElement extends FormItemElement {
 
   protected _render(): void {
     // istanbul ignore else
-    if (this.isConnected && this.objectList) {
-      const objectMap = keyBy(this.objectList, "objectId");
+    if (this.isConnected) {
       const mutableProps = {
         value: this.value,
       };
@@ -233,7 +206,6 @@ export class UserOrUserGroupSelectElement extends FormItemElement {
             name={this.name}
             label={this.label}
             labelTooltip={this.labelTooltip}
-            objectMap={objectMap}
             message={this.message}
             required={this.required}
             validator={this.validator}