+++ bricks/next-builder/src/index.ts
@@ -61,3 +61,4 @@ import "./custom-processors/getWorkbenchRouteTree";
 import "./data-providers/ReplaceSingleThemePage";
 import "./data-providers/PasteBricks";
 import "./shared/components/contract-auto-complete";
+// !Lazy: import "./workbench-brick-context-menu";