+++ bricks/frame-bricks/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.5](https://github.com/easyops-cn/next-basics/compare/@next-bricks/frame-bricks@1.1.4...@next-bricks/frame-bricks@1.1.5) (2022-07-07)
+
+**Note:** Version bump only for package @next-bricks/frame-bricks
+
+
+
+
+
 ## [1.1.4](https://github.com/easyops-cn/next-basics/compare/@next-bricks/frame-bricks@1.1.3...@next-bricks/frame-bricks@1.1.4) (2022-06-29)
 
 **Note:** Version bump only for package @next-bricks/frame-bricks